backend/requirements.txt
Django>=5.0,<6.0
django-rest-framework
django-cors-headers
celery
redis
playwright
playwright-stealth
openai
psycopg2-binary
python-dotenv
requests
dj-database-url
whitenoise
